auth/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-jose>=3.0.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0
httpx>=0.28
email-validator>=2.0
python-multipart>=0.0.9
